requirements.txt
chess==1.11.1
discord.py==2.3.2
python-dotenv==1.0.1
stockfish==3.28.0
Wand==0.6.13
